refactor(backend): modify architecture
