+++ backend/.env.example
@@ -4,6 +4,7 @@ ENV="development" # Options: development, production
 SERVER_HOST="0.0.0.0"
 SERVER_PORT=8080
 SERVER_READ_TIMEOUT=60
+SERVER_URL="${SERVER_HOST}:${SERVER_PORT}"
 
 # Database url for neon:
 DB_URL="postgresql://[user]:[password]@[neon_hostname]/[dbname]?sslmode=require&channel_binding=require"
@@ -12,5 +13,9 @@ DB_URL="postgresql://[user]:[password]@[neon_hostname]/[dbname]?sslmode=require&
 DB_MAX_CONNECTIONS=100
 DB_MAX_IDLE_CONNECTIONS=10
 DB_MAX_LIFETIME_CONNECTIONS=2
+
+# JWT
+JWT_EXPIRES_IN=
+JWT_SECRET=
