--- backend/app/BUSINESS_LOGIC.md
@@ -1,8 +1,0 @@
-# ./app
-
-**Folder with business logic only**. This directory doesn't care about _what database driver you're using_ or _which caching solution your choose_ or any third-party things.
-
-- `./app/controllers` folder for functional controllers (used in routes)
-- `./app/models` folder for describe business models of your project
-- `./app/queries` folder for describe queries for models of your project
-- `./app/validators` folder for describe validators for models fields
