--- backend/pkg/PROJECT_SPECIFIC.md
@@ -1,8 +1,0 @@
-# ./pkg
-
-**Folder with project specific functionality**. This directory contains all the project-specific code tailored only for your business use case, like _configs_, _middleware_, _routes_ or _utils_.
-
-- `./pkg/configs` folder for configuration functions
-- `./pkg/middleware` folder for add middleware (Fiber and yours)
-- `./pkg/routes` folder for describe routes of your project
-- `./pkg/utils` folder with utility functions (server starter, error checker, etc)
