--- backend/platform/PLATFORM_LEVEL.md
@@ -1,7 +1,0 @@
-# ./platform
-
-**Folder with platform-level logic**. This directory contains all the platform-level logic that will build up the actual project, like _setting up the database_ or _cache server instance_ and _storing migrations_.
-
-- `./platform/database` folder with database configuration (by default, PostgreSQL)
-- `./platform/migrations` folder with migration files (used with [golang-migrate/migrate](https://github.com/golang-migrate/migrate) tool)
-- `./platform/seeders` folder with database seeders for populating initial data
